pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alex <42418433+aiexz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {version = "^1.7.4"}
sqlalchemy = "^2.0.19"
pydantic-settings = "^2.0.2"
asyncpg = "^0.28.0"
alembic = "^1.11.1"
argon2-cffi = "^21.3.0"
alembic-postgresql-enum = "^0.1.7"
python-multipart = "^0.0.6"
sentry-sdk = {extras = ["fastapi"], version = "^1.32.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

